+++ nengo/tests/test_synapses.py
@@ -7,8 +7,14 @@ import pytest
 import nengo
 from nengo.exceptions import ValidationError
 from nengo.processes import WhiteSignal
-from nengo.synapses import Alpha, LinearFilter, Lowpass, Synapse, SynapseParam, Triangle
-from nengo.utils.filter_design import cont2discrete
+from nengo.synapses import (
+    Alpha,
+    LinearFilter,
+    Lowpass,
+    Synapse,
+    SynapseParam,
+    Triangle,
+)
 from nengo.utils.testing import signals_allclose
 
 
@@ -45,7 +51,7 @@ def test_direct(Simulator, plt, seed, allclose):
     dt = 1e-3
     a = 0.7
 
-    synapse = LinearFilter([a], [1], analog=False)
+    synapse = LinearFilter(([a], [1]), analog=False)
     t, x, yhat = run_synapse(Simulator, seed, synapse, dt=dt)
     y = synapse.filt(x, dt=dt, y0=0)
 
@@ -69,7 +75,7 @@ def test_alpha(Simulator, plt, seed, allclose):
     num, den = [1], [tau ** 2, 2 * tau, 1]
 
     t, x, yhat = run_synapse(Simulator, seed, Alpha(tau), dt=dt)
-    y = LinearFilter(num, den).filt(x, dt=dt, y0=0)
+    y = LinearFilter((num, den)).filt(x, dt=dt, y0=0)
 
     assert signals_allclose(t, y, yhat, delay=dt, atol=5e-6, plt=plt, allclose=allclose)
 
@@ -107,7 +113,7 @@ def test_decoders(Simulator, plt, seed, allclose):
 
 def test_linearfilter(Simulator, plt, seed, allclose):
     dt = 1e-3
-    synapse = LinearFilter(butter_num, butter_den, analog=False)
+    synapse = LinearFilter((butter_num, butter_den), analog=False)
     t, x, yhat = run_synapse(Simulator, seed, synapse, dt=dt)
     y = synapse.filt(x, dt=dt, y0=0)
 
@@ -116,30 +122,30 @@ def test_linearfilter(Simulator, plt, seed, allclose):
 
 def test_linearfilter_y0(allclose):
     # --- y0 sets initial state correctly for high-order filter
-    synapse = LinearFilter(butter_num, butter_den, analog=False)
+    synapse = LinearFilter((butter_num, butter_den), analog=False)
     v = 9.81
     x = v * np.ones(10)
     assert allclose(synapse.filt(x, y0=v), v)
     assert not allclose(synapse.filt(x, y0=0), v, record_rmse=False)
 
     # --- y0 does not work for high-order synapse when DC gain is zero
-    synapse = LinearFilter([1, 0], [1, 1])
+    synapse = LinearFilter(([1, 0], [1, 1]))
     with pytest.raises(ValidationError, match="Cannot solve for state"):
         synapse.filt(np.ones(10), y0=1)
 
 
 def test_linearfilter_extras(allclose):
     # This filter is just a gain, but caused index errors previously
-    synapse = nengo.LinearFilter([3], [2])
+    synapse = nengo.LinearFilter(([3], [2]))
     assert allclose(synapse.filt([2.0]), 3)
 
     # differentiator should work properly
-    diff = nengo.LinearFilter([1, -1], [1, 0], analog=False)
+    diff = nengo.LinearFilter(([1, -1], [1, 0]), analog=False)
     assert allclose(diff.filt([1.0, -1.0, 2.0], y0=0), [1.0, -2.0, 3.0])
 
     # Filtering an integer array should cast to a float
     x = np.arange(10, dtype=nengo.rc.int_dtype)
-    synapse = nengo.LinearFilter([1], [0.005, 1])
+    synapse = nengo.LinearFilter(([1], [0.005, 1]))
     assert synapse.filt(x).dtype == nengo.rc.float_dtype
 
     # Throw an error if non-float dtype
@@ -210,7 +216,7 @@ def test_lti_lowpass(rng, plt, allclose):
     nt = len(t)
 
     tau = 1e-2
-    lti = LinearFilter([1], [tau, 1])
+    lti = LinearFilter(([1], [tau, 1]))
 
     u = rng.normal(size=(nt, 10))
     x = Lowpass(tau).filt(u, dt=dt)
@@ -227,15 +233,23 @@ def test_linearfilter_combine(rng, allclose):
     nt = 3000
     tau0, tau1 = 0.01, 0.02
     u = rng.normal(size=(nt, 10))
-    x = LinearFilter([1], [tau0 * tau1, tau0 + tau1, 1]).filt(u, y0=0)
+    x = LinearFilter(([], [-1 / tau0, -1 / tau1], 1 / (tau0 * tau1))).filt(u, y0=0)
     y = Lowpass(tau0).combine(Lowpass(tau1)).filt(u, y0=0)
     assert allclose(x, y)
 
-    with pytest.raises(ValidationError, match="other LinearFilters"):
+    with pytest.raises(ValidationError, match="other LinearFilter"):
         Lowpass(0.1).combine(Triangle(0.01))
 
     with pytest.raises(ValidationError, match="analog and digital"):
-        Lowpass(0.1).combine(LinearFilter([1], [1], analog=False))
+        Lowpass(0.1).combine(LinearFilter(([1], [1]), analog=False))
+
+
+def test_linearfilter_validation_errors():
+    with pytest.raises(ValidationError, match="can only have one input"):
+        LinearFilter((-np.ones((2, 2)), np.ones((2, 2)), None, None))
+
+    with pytest.raises(ValidationError, match="can only have one output"):
+        LinearFilter((-np.ones((2, 2)), np.ones((2, 1)), np.ones((2, 2)), None))
 
 
 def test_combined_delay(Simulator, allclose):
@@ -246,15 +260,15 @@ def test_combined_delay(Simulator, allclose):
     dt = 0.001
 
     sys1 = nengo.Lowpass(tau)
-    (num,), den, _ = cont2discrete((sys1.num, sys1.den), dt=dt)
-    sys2 = nengo.LinearFilter(np.poly1d(num) ** 2, np.poly1d(den) ** 2, analog=False)
+    sys1d = sys1.discretize(dt)
+    sys2d = sys1d.combine(sys1d)
 
     with nengo.Network() as model:
         u = nengo.Node(1)
         x = nengo.Node(size_in=1)
         nengo.Connection(u, x, synapse=sys1)
         p1 = nengo.Probe(x, synapse=sys1)
-        p2 = nengo.Probe(u, synapse=sys2)
+        p2 = nengo.Probe(u, synapse=sys2d)
 
     with Simulator(model, dt=dt) as sim:
         sim.run_steps(nt)
@@ -291,9 +305,9 @@ def test_synapseparam():
 
 def test_frozen():
     """Test attributes inherited from FrozenObject"""
-    a = LinearFilter([1], [0.04, 1])
-    b = LinearFilter([1], [0.04, 1])
-    c = LinearFilter([1], [0.04, 1.1])
+    a = LinearFilter(([1], [0.04, 1]))
+    b = LinearFilter(([1], [0.04, 1]))
+    c = LinearFilter(([1], [0.04, 1.1]))
 
     assert hash(a) == hash(a)
     assert hash(b) == hash(b)
@@ -307,7 +321,7 @@ def test_frozen():
     assert hash(b) != hash(c)  # not guaranteed, but highly likely
 
     with pytest.raises((ValueError, RuntimeError)):
-        a.den[0] = 9
+        a.A[0, 0] = 9
 
 
 def test_argreprs():
@@ -317,9 +331,11 @@ def test_argreprs():
     def check_repr(obj):
         assert eval(repr(obj)) == obj
 
-    check_init_args(LinearFilter, ["num", "den", "analog", "method"])
-    check_repr(LinearFilter([1, 2], [3, 4]))
-    check_repr(LinearFilter([1, 2], [3, 4], analog=False))
+    check_init_args(
+        LinearFilter, ["sys", "den", "analog", "method", "x0", "default_dt"]
+    )
+    check_repr(LinearFilter(([1, 2], [3, 4])))
+    check_repr(LinearFilter(([1, 2], [3, 4]), analog=False))
 
     check_init_args(Lowpass, ["tau"])
     check_repr(Lowpass(0.3))