+++ CHANGES.rst
@@ -54,6 +54,11 @@ Release history
   read-only, to prevent users from accidentally overwriting them with their own data.
   (`#1545 <https://github.com/nengo/nengo/issues/1545>`__,
   `#1608 <https://github.com/nengo/nengo/pull/1608>`__)
+- ``LinearFilter`` now accepts system descriptions in state-space (``(A, B, C, D)``) or
+  zero-pole-gain (``(z, p, k)``) form, in addition to transfer function (``(num, den)``)
+  form. If passing systems in transfer function form, they should be passed as a
+  ``(numerator, denominator)`` tuple; the separate ``den`` argument is deprecated.
+  (`#1611 <https://github.com/nengo/nengo/pull/1611>`__)
 
 **Fixed**